+++ packages/agility_redux_bloc/example/lib/counter_bloc/counter_page/counter_page.dart
@@ -25,6 +25,16 @@ class CounterPage extends BlocPageRoute {
 
   @override
   Widget build(BuildContext context) {
-    return CounterWidget();
+    return AppNavigatorContainer(
+      child: CounterWidget(),
+      background: Column(
+        children: [
+          Container(
+            height: 56.0,
+            color: Colors.blue,
+          ),
+        ],
+      ),
+    );
   }
 }